requirements.txt
hydra-core
torch
transformers
tqdm
openai
nltk
scikit-learn